# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErasureBody": {
        "additionalProperties": false,
        "properties": {
          "driver_id": {
            "minimum": 1.0,
            "title": "Driver Id",
            "type": "integer"
          }
        },
        "required": [
          "driver_id"
        ],
        "title": "ErasureBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\nRosetta translates vehicle telemetry from many car makers into one canonical event.\n\n* **Authentication** OAuth2 bearer tokens (JWT). Get one from `POST /api/v1/auth/token`.\n* **Errors** RFC 9457 problem details (`application/problem+json`).\n* **Pagination** keyset cursors: pass `next_cursor` back as `cursor`. No offsets.\n* **Rate limits** per caller. A 429 carries `Retry-After`.\n* **Versioning** the path carries the major version. Additive changes do not bump it.\n",
    "title": "Rosetta API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "operationId": "audit_log_api_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "actor_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(user|agent|service|system|anonymous)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Kind"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "pattern": "^[a-z_.]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "The audit trail, newest first",
        "tags": [
          "compliance"
        ]
      }
    },
    "/api/v1/audit/verify": {
      "get": {
        "operationId": "audit_verify_api_v1_audit_verify_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Recompute the hash chain and report whether it is intact",
        "tags": [
          "compliance"
        ]
      }
    },
...
    "/api/v1/compliance/erasure": {
      "get": {
        "operationId": "erasures_api_v1_compliance_erasure_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Erasures Api V1 Compliance Erasure Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Erasure requests and their evidence",
        "tags": [
          "compliance"
        ]
      },
      "post": {
        "operationId": "erase_api_v1_compliance_erasure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ErasureBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Erase a driver's personal data and prove it",
        "tags": [
          "compliance"
        ]
      }
    },
    "/api/v1/compliance/policy": {
      "get": {
        "operationId": "policy_api_v1_compliance_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "What is stored, where, for how long, and who may see it",
        "tags": [
          "compliance"
        ]
      }
    },
...
    "/api/v1/drivers": {
      "get": {
        "operationId": "drivers_api_v1_drivers_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Drivers of the caller's tenant (personal data)",
        "tags": [
          "compliance"
        ]
      }
    },
...
  }
}